Updated StateChange fields and stateChange for transitions to same state.
Changed StateChange from/to fields to be exported. The StateChange struct is
public, but its fields weren't, meaning ChanHandler was useless to know what
the transition was for.

Also changed node.stateChange to not trigger a state change notification if
the new state is the same as the current state. It was noticed that if a
client is in the candidate state, then it would send the notification each
time the voting timeout expired even though it wasn't actually in a different
state.