Update Travis + Fixed flapping tests
Was getting failures with `-race` for heartbeat tests.
The util function sendAndWait() just ensures that the heartbeat
has been dequeued. Change in state and/or leader is not guaranteed
at this point. Added waitForLeader and waitForState to wait for
a certain amount of time for the expected leader/state. Used these
in various places, may need to use it in some other places if we
get other tests failures.