Ensure StateChange are posted in order to Handler
This is another approach for solving issue #11 that was addressed
with PR #12. The advantage of this approach over #12 is that there
will be atmost 1 go-routine per type of event (StateChange and error).

Added a test that demonstrates that not receiving from state change
and error channels do not block the Node.

Also fixed a flapping test.

Resolves #11